requirements.txt
asgiref==3.2.10
astroid==2.4.2
cairocffi===file-.cairocffi-VERSION
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.0.7
django-cors-headers==3.4.0
django-crontab==0.7.1
django-filter==2.3.0
django-mathfilters==1.0.0
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
djangorestframework==3.11.0
EasyProcess==0.3
fpdf==1.7.2
html5lib==1.1
idna==2.10
importlib-metadata==1.6.1
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.19.1
pandas==1.1.2
pdfkit==0.6.1
pep8==1.7.1
Pillow==7.1.2
pkg-resources==0.0.0
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
PyMySQL==0.9.3
pyodbc==4.0.30
pyotp==2.4.0
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
PyVirtualDisplay==1.3.2
reportlab==3.5.47
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
toml==0.10.1
twilio==6.45.1
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.10
WeasyPrint==51
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.4
xvfbwrapper==0.2.9
zipp==3.1.0
